Removed sharding support
After further research, I realized the integrated sharding support in discord.py is intended for handling multiple guilds (servers), not for multiple bots on the same server. After discussing this with a friend, he thinks that built-in sharding is a bit overkill compared to just restarting the bot on crash: mostly comparable downtime, and less management to deal with failover.